add service locator
This is meant to be used to located services, of which there are only
one. Examples of such services might be the MeshManager, or audio
engine. This was made as a hack to get Wren entity scripting working,
as it was a decent way obtaining an access point into a free
function.